Creacion de Component.css
