+++ src/ui/components/ComponentBox/ComponentBox.css
@@ -1,0 +1,39 @@
+
+.component-box {
+    position: relative;
+    width: 100%;
+    height: 100%;
+  }
+  
+  .circle {
+    width: 150px;
+    height: 150px;
+    border-radius: 50%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+  }
+  
+  .number {
+    font-size: 30px;
+    font-weight: bold;
+    margin-bottom: 5px;
+  }
+  
+  .inner-text {
+    font-size: 16px;
+  }
+  
+  .outer-text {
+    font-size: 16px;
+    position: absolute;
+    top: 5px;
+    left: 5px;
+  }
+
